requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pyrebase4==4.5.0
pydantic==1.10.7
python-dotenv==0.20.0
